refactor(BaggageStepper): refactor to ts
