+++ packages/orbit-components/src/BaggageStepper/Stepper/consts.ts
@@ -1,3 +1,2 @@
-// @flow
 export const ICON_SIZE = "10px";
 export const BUTTON_SIZE = "20px";
