+++ packages/orbit-components/src/BaggageStepper/__tests__/index.test.tsx
@@ -1,4 +1,3 @@
-// @flow
 import * as React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
